+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 * scoreboard
 * padding pre
 
+# 0.6.3-SNAPSHOT
+
+- resources/users.{json,txt}
+- refactored workspace.clj
+- added system/restart-system - (stop-system) (start-system) is better than clj-reload(?)
+- updated system/start-server, system/stop-server - only when available, doing the jobs.
 
 # 0.6.2 (2025-11-27)